Refactor into common
