+++ common/analysis.py
