+++ common/annotation_transformer.py
@@ -1,6 +1,6 @@
 import ast
 
-from .analysis import get_id
+from common.analysis import get_id
 
 def add_annotation_flags(node):
     return AnnotationTransformer().visit(node)
