+++ pyrs/tests/test_scope.py
@@ -1,6 +1,6 @@
 import ast
-from pyrs.scope import add_scope_context
-from pyrs.context import add_variable_context
+from common.scope import add_scope_context
+from common.context import add_variable_context
 
 
 def parse(*args):
